put safety check to standalone step
